Modal component creation and component abstraction
